requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
dataclasses-json==0.5.14
faiss-cpu==1.7.4
fastapi==0.101.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
langchain==0.0.209
langchainplus-sdk==0.0.20
langsmith==0.0.20
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==1.4.1
pdfminer==20191125
pdfminer.six==20221105
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.2
tqdm==4.66.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
python-dotenv
tiktoken
langdetect
